changed passiord

+++ C_Code/LARK_Libraries/entity_functions.c
@@ -56,7 +56,7 @@ void UseKey(Transform *key){
 
 void TriggerTrap(Transform *t){
     if(t->level == LEVEL_LOADED && (int)t->position.x == (int)PLAYER->position.x && (int)t->position.y == (int)PLAYER->position.y){
-        float trapDMG = -15;
+        int trapDMG = -15;
         PLAYER->OnUse(&trapDMG);
         sprintf(terminalOutput, "\r\nPLAYER STEPPED ON [%s], [%d] damage was done\r\nUse './stats.sh' to see current health", t->name, -15);
     }
